- dodane nowe zmienne - ujednolicenie nazw zmiennych opisujących studiowanie (sufiks "_st" implikuje bycie stała, dla zmiennych okienkowych przy dowolnej kategorii zatrudnienia sufiks "_x_s") - korekta wyliczania zmiennych "if_x_s" oraz "if_x_strprg" (dawniej "if_st" i "if_stprg"), które muszą być wyliczane na poziomie {id, okres}, a nie, jak do tej pory {id, okres, pracodawca, tytuł ubezpieczenia} - zmiana rodzaju dużej ilości zmiennych z numeric na integer (co powinno dać kilka procent szybsze obliczenia)

+++ R/data2okres.R
@@ -8,5 +8,5 @@
 #' @return numeric liczba miesięcy jako ROK*12 + MIESIĄC
 #' @export
 data2okres = function(data){
-  return(as.numeric(sub('-.*$', '', data)) * 12 + as.numeric(sub('^[0-9]+-([0-9]+).*$', '\\1', data)))
+  return(as.integer(sub('-.*$', '', data)) * 12L + as.integer(sub('^[0-9]+-([0-9]+).*$', '\\1', data)))
 }
